Add files via upload
This is the first running version of the program that takes user input to return a snow water equivalent.  Possible routes from here may include using satellite data to determine actual angles of elevation of the signal received at the GPS and/or plotting snow water equivalent values for a certain time period from data in an excel spreadsheet.